Complete 'primitive' module
